ts: get tests running/passing
A number of conventions seem to have changed, causing a fair bit of breakage:
  - `Data_*` was deprecated in favor of `TypedData_*`
  - `ossl_obj2bio` takes a `VALUE*` instead of `VALUE` now
  - `time_to_time_t()` was removed
